Move signals to their own files
